fix: pass return values through play_round
Change the computerChoice variable inside of getComputerChoice to a
parameter. Return computerChoice and humanChoice. Remove the call for
getComputerChoice and getHumanChoice, then assign the computerChoice
and humanChoice parameters inside of playRound to the corresponding
functions.

This sucessfully passes the original values of
computerChoice and humanChoice to playRound. Also, remove console.log
inside of getComputerChoice that displays the value of computerChoice
because new console.logs were made inside of playRound to do the same
thing.