docs: make small formatting changes to logs
Remove console.log that displays random number's value because it isn't
needed anymore and is cluttering the console. For the if statement that
is logging computerChoice and humanChoice, add Computer: and Human:
respectively to the console.log. This adds clarity to who is choosing
what. In the if statement that logs computerScore and humanScore,
change the order to display the humanScore first. It looks slightly
better and provides a bit more clarity.