feat: make the round outcome display on screen
Inside of all the if statements that change the round outcome, create
paragraphs. Give these paragraphs the text content of console.log that
matches the winner, loser, humanChoice, and computerChoice. Make a
reference to the output div then use that reference to append the
paragraphs so that they actually appear on screen.