feat: display a game victory or loss message
If humanScore is equal to five display a victory message. If
computerScore is equal to five display a loss message.

Add a linebreak to the end of the if statements to space out the the
victory or loss message from the score and outcomeMsg. This maintains
readability.

Move the if statement that resets humanScore and computerScore to zero
after it reaches five below the if statements containing gameOutcome.
Otherwise, humanScore and computerScore will reset too early, causing
gameOutcome to never display a victory or loss message.