fix: check for invalid input from computer
The else statement was checking for a draw and one else if statement
was checking to see if humanChoice equalled something other than
rock, paper, or scissors. There was nothing checking to see if the
computer's input was invalid, which has been found to be true on rare
occasions.

Swap the code inside of the else and last else if statement, then
make else if check for draws. The else statement looks for everything
else, which at this point should exclusivly be invalid input. It then
displays a message letting the player know that the input was invalid.