fix: replace math-random in get-computer-choice
In getComputerChoice replace Math.random with getRandomInt(0, 2).
Afterwards, rewrite the if statements that assign a value to
computerChoice so that they take the integers 0, 1, and 2 instead of
specific decimals. This assigns a unique number to rock, paper, and
scissors. It also makes the code easier to understand.

Do this because the code that was previously used had a small chance
of giving computerChoice the value undefined, which is not at all
ideal.